# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Immobilière",
        "description": "API pour la gestion des biens immobiliers",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tenant/{locataireId}/paiements": {
            "get": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Lister les paiements d'un locataire",
                "operationId": "d8e9f4cd8b406f507606012c1d0e4848",
                "parameters": [
                    {
                        "name": "locataireId",
                        "in": "path",
                        "description": "ID du locataire",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des paiements récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "locataire": {
                                                    "$ref": "#/components/schemas/LocataireWithPaiements"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Paiements récupérés avec succès"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Locataire non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Locataire non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Enregistrer un nouveau paiement",
                "operationId": "3bbf431b6638c32ef23fbe410c1eee19",
                "parameters": [
                    {
                        "name": "locataireId",
                        "in": "path",
                        "description": "ID du locataire",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "mois_couvert",
                                    "methode_paiement"
                                ],
                                "properties": {
                                    "mois_couvert": {
                                        "description": "Mois couvert par le paiement (format YYYY-MM)",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-12"
                                    },
                                    "methode_paiement": {
                                        "description": "Méthode de paiement",
                                        "type": "string",
                                        "enum": [
                                            "mobile_money",
                                            "virement"
                                        ],
                                        "example": "mobile_money"
                                    },
                                    "transaction_id": {
                                        "description": "ID de transaction (requis pour mobile money)",
                                        "type": "string",
                                        "example": "MM_123456789"
                                    },
                                    "proof_file": {
                                        "description": "Fichier de preuve (requis pour virement bancaire)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Paiement enregistré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Paiement enregistré avec succès pour le mois de Décembre 2023"
                                        },
                                        "paiement": {
                                            "$ref": "#/components/schemas/Paiement"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Locataire non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Locataire non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Paiement déjà existant",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Paiement déjà enregistré pour Décembre 2023"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Le loyer pour Décembre 2023 a déjà été payé."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "mois_couvert": [
                                                    "Le champ mois couvert est requis."
                                                ],
                                                "methode_paiement": [
                                                    "La méthode de paiement sélectionnée est invalide."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur lors de l'enregistrement du paiement"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Message d'erreur détaillé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tenant/paiements/{id}": {
            "get": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Afficher les détails d'un paiement",
                "operationId": "8d476501be5f66267621b70892b589b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du paiement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du paiement récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Détails du paiement récupérés avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PaiementDetails"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paiement non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Paiement non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LocataireWithPaiements": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Locataire"
                    },
                    {
                        "properties": {
                            "paiements": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Paiement"
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "Paiement": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "reference": {
                        "type": "string",
                        "example": "PAY-12345"
                    },
                    "montant": {
                        "type": "number",
                        "format": "float",
                        "example": 500
                    },
                    "date_paiement": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-26T10:00:00Z"
                    },
                    "mois_couvert": {
                        "type": "string",
                        "example": "2023-10"
                    },
                    "statut": {
                        "type": "string",
                        "enum": [
                            "payé",
                            "En attente"
                        ],
                        "example": "payé"
                    }
                },
                "type": "object"
            },
            "PaiementDetails": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Paiement"
                    },
                    {
                        "properties": {
                            "locataire": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "prenom": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            },
                            "bien": {
                                "properties": {
                                    "commune": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "Locataire": {
                "title": "Locataire",
                "description": "Modèle de locataire",
                "required": [
                    "name",
                    "email",
                    "password",
                    "contact"
                ],
                "properties": {
                    "id": {
                        "description": "ID du locataire",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Nom du locataire",
                        "type": "string"
                    },
                    "prenom": {
                        "description": "Prénom du locataire",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email du locataire",
                        "type": "string",
                        "format": "email"
                    },
                    "contact": {
                        "description": "Contact du locataire",
                        "type": "string"
                    },
                    "piece": {
                        "description": "Type de pièce d'identité",
                        "type": "string"
                    },
                    "adresse": {
                        "description": "Adresse du locataire",
                        "type": "string"
                    },
                    "profession": {
                        "description": "Profession du locataire",
                        "type": "string"
                    },
                    "status": {
                        "description": "Statut du locataire",
                        "type": "string",
                        "enum": [
                            "Actif",
                            "Inactif"
                        ]
                    },
                    "agence_id": {
                        "description": "ID de l'agence associée",
                        "type": "integer"
                    },
                    "proprietaire_id": {
                        "description": "ID du propriétaire associé",
                        "type": "integer"
                    },
                    "contrat_id": {
                        "description": "ID du contrat associé",
                        "type": "integer"
                    },
                    "bien_id": {
                        "description": "ID du bien associé",
                        "type": "integer"
                    },
                    "image1": {
                        "description": "Image 1 du locataire",
                        "type": "string",
                        "format": "binary"
                    },
                    "image2": {
                        "description": "Image 2 du locataire",
                        "type": "string",
                        "format": "binary"
                    },
                    "image3": {
                        "description": "Image 3 du locataire",
                        "type": "string",
                        "format": "binary"
                    },
                    "image4": {
                        "description": "Image 4 du locataire",
                        "type": "string",
                        "format": "binary"
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}